fix: error when element deleted after css update (#2553)
* fix: revert PR template because issues not closing

* fix: css.cy test and remove it from ignore

* fix: get rid of redundant requests to server

* fix: css.cy spec instability

* fix: avoid re-renders
